Cache Mac error messages at every point control passes between OpenSolverMain and a class module.
Excel 2011 loses all error information when control passes from a class module back into a normal module, so we need to save the messages before this happens.

All error messages in tests are displaying correctly on mac and windows, but it is possible we need to add caching at other points.

Fixes #13