Mends the Italian language bug, where a difference between the regional and User-set decimal seperators was causing a localisation routine to fail.
Also introduces a temporary workbook for these localisation routines to use, instead of the old Cell (1,1) of sheet 1 in the users workbook, which was being overwritten. (I confirmed this, Andrew). The book is hidden from the user, and is deleted after successful execution or on any error that is raised to the top level. It does linger after pressing debug as a dev, though, so I'll post an issue to resolve this.

Fixes #43
Fixes #44