Related to #49, adds a MBox function in OpenSolverModule that will replace the use of msgbox throughout OpenSolver.
Need to locate the hand cursor used in the hyperlink on the about opensolver form (I can see how to implement it, but can't find the icon file), and also need to determine when and where we should link to the help page. All the time (even when it's due to an error in their model), or only when they've hit a genuine error? Probably the lattter, but the definition of genuine could use some thought.