Fix issue with Gurobi in 64-bit Excel.
Gurobi can return small numbers (e.g. 1e-323) which are too small for Excel to understand and so the conversion from string to float fails.

As a solution we can strip the last digit in the exponent (still giving a very small number!) to get a string that Excel can understand.